+++ src/components/rangeUsePlanPage/pageForAH/index.js
@@ -25,7 +25,7 @@ import AdditionalRequirements from '../additionalRequirements'
 import ManagementConsiderations from '../managementConsiderations'
 import EditableManagementConsiderations from '../editableManagementConsiderations'
 import { defaultProps, propTypes } from './props'
-import ActionBtns from './ActionBtns'
+import ActionBtns from '../ActionBtns'
 import PlanSubmissionModal from './SubmissionModal'
 import AHSignatureModal from './AHSignatureModal'
 import AmendmentSubmissionModal from './AmendmentSubmissionModal'
